feat : 폴더 구조 변경, airflow ecs 환경
- airflow 폴더 생성하여 폴더 구조 변경
- 원래 dags/plugins 폴더를 한 단계 상위로 변경
- airflow ecs 환경 cd (ecr푸시 확인하는 중)
- airflow 폴더 안에 있는 requirements.txt의 내용을 루트 폴더의 requirements.txt에도 추가함(테스트)
