+++ airflow/requirements.txt
@@ -1,0 +1,4 @@
+apache-airflow[docker,amazon]
+typing
+s3fs
+pandas
