--- data/news_article/news_article_2023-08-19.csv
@@ -1,1 +1,0 @@
-
